Moved user mode events to external folder
