+++ tracee-ebpf/external/userevents/system_info.go
