+++ tracee-ebpf/external/userevents/system_info_test.go
